Agregar modal de confirmacion

+++ src/app/products/products.component.ts
@@ -1,18 +1,24 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from '../models/product.model';
 import { ProductService } from '../services/product.service';
+import { State } from '../app.enums';
 
 @Component({
     selector: 'app-products',
     templateUrl: './products.component.html'
 })
 export class ProductsComponent implements OnInit {
+    question: string;
+    isActive: boolean;
     products: Product[];
+    productToDelete: Product;
 
     constructor(private productService: ProductService) {}
 
     ngOnInit() {
         this.products = [];
+        this.isActive = false;
+        this.question = State.empty
         this.getProducts();        
     }
 
@@ -24,14 +30,27 @@ export class ProductsComponent implements OnInit {
     }
 
     deleteProduct(product: Product) {
-        const { id } = product;
+        this.question = `
+            ¿Esta seguro que desea eliminar el producto <b>"${product.name}"</b>?
+        `;
+        this.isActive = true;
+        this.productToDelete = product;
+    }
 
-        this.productService.deleteProduct(id)
-            .subscribe(response => {
-                this.products = this.products
-                    .filter(prod => prod.id !== id);
-                // this.getProducts();
-            });
+    onResponse(confirmation: boolean) {
+        if (confirmation) {
+            const { id } = this.productToDelete;
+
+            this.productService.deleteProduct(id)
+                .subscribe(response => {
+                    this.products = this.products
+                        .filter(prod => prod.id !== id);
+                    this.isActive = false;
+                });
+        } else {
+            this.isActive = false;
+        }
+
     }
 
     changeState(product: Product) {
